+++ src/main/java/org/nette/latte/completion/resolvers/ControlResolver.java
@@ -1,0 +1,72 @@
+package org.nette.latte.completion.resolvers;
+
+import com.intellij.psi.PsiElement;
+import com.jetbrains.php.lang.psi.elements.Method;
+import com.jetbrains.php.lang.psi.elements.PhpClass;
+import org.apache.commons.lang3.StringUtils;
+import org.jetbrains.annotations.Nullable;
+import org.nette.latte.psi.LatteFile;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class ControlResolver extends PresenterResolver {
+    //private final HashMap<String, @Nullable PsiElement> cache = new HashMap<>();
+
+    public ControlResolver(LatteFile file) {
+        super(file);
+    }
+
+    public void reset() {
+        //cache.clear();
+    }
+
+    public @Nullable PsiElement resolveComponent(String component) {
+        String key = component;
+        /*if (cache.containsKey(key)) {
+            return cache.get(key);
+        }*/
+
+        PsiElement result = calculateComponent(component);
+        //cache.put(key, result);
+        return result;
+    }
+
+    private @Nullable PsiElement calculateComponent(String component) {
+        List<String> presenterNames = guessPresenterNames();
+        List<PhpClass> matchingPresenters = getMatchingPresenters(presenterNames, false);
+
+        if (matchingPresenters.isEmpty()) {
+            return null;
+        }
+
+        for (PhpClass presenterClass : matchingPresenters) {
+            Method method = findMethod(presenterClass, List.of("createComponent" + StringUtils.capitalize(component)));
+            if (method != null && (!method.getName().equals("startup") || method.getClass().getName().equals(presenterClass.getName()))) {
+                return method;
+            }
+        }
+
+        return null;
+    }
+
+    public List<Method> getComponents() {
+        List<String> presenterNames = guessPresenterNames();
+        List<PhpClass> matchingPresenters = getMatchingPresenters(presenterNames, false);
+
+        if (matchingPresenters.isEmpty()) {
+            return new ArrayList<>();
+        }
+
+        PhpClass presenterClass = matchingPresenters.get(0);
+
+        List<Method> components = new ArrayList<>();
+        for (Method method : presenterClass.getMethods()) {
+            if (method.getName().startsWith("createComponent")) {
+                components.add(method);
+            }
+        }
+
+        return components;
+    }
+}